requirements.txt
tkcalendar
mysql-connector-python
bcrypt
python-dotenv
openpyxl
pillow
requests
pytest
